NexusTextureFactory-Tezut/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-texture-factory-tezut"
version = "0.1.0"
description = "Wheel-authoritative Python CLI for Nexus Texture Factory Tezut"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=11.0.0",
]

[project.scripts]
nexus-tezut = "nexus_texture_factory_tezut.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

